fix(ServiceDetailPage): move notification refresh logic to a separate utility file
